Database Fixes

+++ gamemode/backend/libs/netlogging/sv_netmessagelogger.lua
@@ -1,15 +1,11 @@
 util.AddNetworkString("net_ReceiveLogs")
 util.AddNetworkString("net_RequestLogs")
 local logs = {}
-
 local function Init()
-    if not file.Exists("netmessagelogsdir", "DATA") then
-        file.CreateDir("netmessagelogsdir")
-    end
+    if not file.Exists("netmessagelogsdir", "DATA") then file.CreateDir("netmessagelogsdir") end
 end
 
 Init()
-
 local function loadLogs()
     local data = file.Read("netmessagelogs.txt")
     if not data then return end
@@ -22,7 +18,6 @@ local function saveLogs()
 end
 
 loadLogs()
-
 function net.Incoming(len, ply)
     local i = net.ReadHeader()
     local strName = util.NetworkIDToString(i)
@@ -32,7 +27,6 @@ function net.Incoming(len, ply)
     len = len - 16
     local debugInfo = debug.getinfo(func, "S")
     local timeString = os.date("%d/%m/%Y - %H:%M:%S", os.time())
-
     local logEntry = {
         ["time"] = timeString,
         ["name"] = strName,
@@ -47,11 +41,7 @@ function net.Incoming(len, ply)
     if #logs >= 5000 then
         local files = file.Find("netmessagelogsdir/*.txt", "DATA", "dateasc")
         local count = #files + 1
-
-        if count > 50 then
-            file.Delete("netmessagelogsdir/" .. files[1])
-        end
-
+        if count > 50 then file.Delete("netmessagelogsdir/" .. files[1]) end
         file.Write("netmessagelogsdir/" .. os.date("%d-%m-%Y_%H-%M-%S", os.time()) .. "_netmessagelogs.txt", util.TableToJSON(logs))
         logs = {}
     end
@@ -61,13 +51,11 @@ function net.Incoming(len, ply)
 end
 
 hook.Add("ShutDown", "SaveNetMessageLogs", saveLogs)
-
 function sendData(page, ply)
     if not ply:IsSuperAdmin() then return end
     local dataTable = {}
     local start = ((page - 1) * 128) + 1
     local finish = math.min(start + 127, #logs)
-
     for i = start, finish do
         table.insert(dataTable, logs[i])
     end
@@ -81,7 +69,10 @@ function sendData(page, ply)
     net.Send(ply)
 end
 
-net.Receive("net_RequestLogs", function(len, ply)
-    local page = net.ReadInt(32)
-    sendData(page, ply)
-end)
+net.Receive(
+    "net_RequestLogs",
+    function(len, ply)
+        local page = net.ReadInt(32)
+        sendData(page, ply)
+    end
+)
